requirements.txt
asgiref==3.8.1
captcha==0.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
Django==5.1.1
django-allauth==65.0.2
django-allauth-2fa==0.11.1
django-otp==1.5.4
django-recaptcha==4.0.0
fido2==1.1.3
gunicorn==20.1.0
idna==3.10
jwt==1.3.1
pillow==10.4.0
pycparser==2.22
qrcode==8.0
requests==2.32.3
setuptools==75.1.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.4.0
